### 2025-01-08 #7636 @editorialbot
[PRE REVIEW]: Groundhog - A general-purpose geotechnical Python package
**Submitting author:** @snakesonabrain (<a href="http://orcid.org/0000-0002-1844-8634">Bruno Stuyts</a>)
**Repository:** https://github.com/snakesonabrain/groundhog
**Branch with paper.md** (empty if default branch): JOSS-paper
**Version:** v0.13.0
