+++ groundhog/soildynamics/cptliquefaction.py
@@ -1,0 +1,907 @@
+__author__ = "Wouter Karreman & Ping Li"
+
+# Native Python packages
+import math
+
+# 3rd party packages
+import numpy as np
+
+# Project imports
+from groundhog.general.validation import Validator
+
+FOS_LIQUEFACTION = {
+    "sigma_vo": {"type": "float", "min_value": 0.0, "max_value": None},
+    "sigma_vo_eff": {"type": "float", "min_value": 0.0, "max_value": None},
+    "CRR": {"type": "float", "min_value": 0.0, "max_value": None},
+    "CSR": {"type": "float", "min_value": 0.0, "max_value": None},
+    "MSF": {"type": "float", "min_value": 0.0, "max_value": None},
+    "K_sigma": {"type": "float", "min_value": 0.0, "max_value": None},
+}
+
+FOS_LIQUEFACTION_ERRORRETURN = {"FoS_liq [-]": np.nan}
+
+
+@Validator(FOS_LIQUEFACTION, FOS_LIQUEFACTION_ERRORRETURN)
+def fos_liquefaction(sigma_vo, sigma_vo_eff, CRR, CSR, MSF, K_sigma, **kwargs):
+    """
+    Calculates the Factor of Safety (FoS) for liquefaction assessment
+
+    Args:
+        sigma_vo [kPa]: Total vertical stress at the depth
+        sigma_vo_eff [kPa]: Effective vertical stress at the depth
+        CRR [-]: Cyclic resistance ratio
+        CSR [-]: cyclic stress ratio
+        MSF [-]: Magnitude scaling factor
+        K_sigma [-]: Overburden correction factor
+
+    Returns:
+        FoS_liq [-]: Factor of Safety (FoS) for liquefaction assessment
+    """
+    if sigma_vo_eff == sigma_vo:
+        FoS = 5
+    else:
+        FoS = min((CRR / CSR) * MSF * K_sigma, 5)
+
+    return {
+        "FoS_liq [-]": FoS,
+    }
+
+
+CSR_ROBERTSON_CABAL_2022 = {
+    "sigma_vo": {"type": "float", "min_value": 0.0, "max_value": None},
+    "sigma_vo_eff": {"type": "float", "min_value": 0.0, "max_value": None},
+    "depth": {"type": "float", "min_value": 0.0, "max_value": None},
+    "magnitude": {"type": "float", "min_value": 5.5, "max_value": 8.5},
+    "acceleration": {"type": "float", "min_value": 0.0, "max_value": None},
+    "MSF_userdefined": {
+        "type": "bool",
+    },
+}
+
+CSR_ROBERTSON_CABAL_2022_ERRORRETURN = {
+    "MSF [-]": np.nan,
+    "CSR [-]": np.nan,
+}
+
+
+@Validator(CSR_ROBERTSON_CABAL_2022, CSR_ROBERTSON_CABAL_2022_ERRORRETURN)
+def csr_robertson_cabal_2022(
+    sigma_vo,
+    sigma_vo_eff,
+    depth,
+    magnitude,
+    acceleration,
+    MSF_userdefined=False,
+    **kwargs
+):
+    """
+    Calculates CSR following Robertson and Cabal (2022)
+
+    Args:
+        sigma_vo [kPa]: Total vertical stress at the depth
+        sigma_vo_eff [kPa]: Effective vertical stress at the depth
+        depth [m]: Depth below mudline
+        magnitude [-]: Earthquake magnitude
+        acceleration [-]: Maximum horizontal acceleration at the soil surface
[…]
