### 2025-02-28 #7636 @snakesonabrain
@kthyng Apologies for being a bit late to the game here. I'm picking this up now. I think I will just remove the section with examples as these are included in the tutorials.
