+++ JOSS/paper.md
@@ -60,71 +60,6 @@ The geotechnical parameter correlations and geotechnical workflow automation too
 
 A dedicated two-day training course has also been set up to teach Python and `Groundhog` to geotechnical engineers seeking to get started with scripting. This course has been delivered to Systra in Dubai (April 2024) and RINA in Milan (October 2024).
 
-# Selected examples
-
-## Parameter validation
-
-The parameter validation functionality can be demonstrated using the function `relativedensity_sand_jamiolkowski` which is used to estimate the relative density of sand from CPT data. The function takes cone tip resistance (`qc`), vertical effective stress (`sigma_vo_eff`) and the coefficient of lateral earth pressure (`k0`) as inputs. The function returns two unitless outputs as a Python dictionary; the relative density for dry sand (`Dr dry [-]`) and the relative density for saturated conditions (`Dr sat [-]`).
-
-The function was developed based on calibration chamber tests with vertical effective stresses for these tests varying between 50kPa and 400kPa. Hence, only vertical effective stresses in that range are allowed. When a user attempts to execute the function for a vertical effective stress of 10kPa, the function will return `nan` for both outputs.
-
-However, the user can override the validation range. By appending `__min` or `__max` to the argument for which the validation range needs overriding, the default validation range is adjusted. By providing `sigma_vo_eff__min=10` as an additional argument, the function will return numerical values for the relative density even though the selected value of vertical effective stress is lower than 50kPa. The overriding of the validation range is explicitly visible in the function call, a minimum vertical effective stress of 10kPa is allowed. The outputs are not realistic (relative density cannot be greater than 1), but the reader of the code can directly connect this to the vertical effective stress which is outside the default validation range.
-
-```python
->>> from groundhog.siteinvestigation.insitutests.pcpt_correlations import \
-    relativedensity_sand_jamiolkowski
->>> relativedensity_sand_jamiolkowski(
-    qc=20, sigma_vo_eff=10, k0=0.8)
-{'Dr dry [-]': nan, 'Dr sat [-]': nan}
->>> relativedensity_sand_jamiolkowski(
-    qc=20, sigma_vo_eff=10, k0=0.8, sigma_vo_eff__min=10)
-{'Dr dry [-]': 1.083398634171661, 'Dr sat [-]': 1.252988806617474}
-```
-
-## Plotting of geotechnical data and stratigraphy
-
-Plotting geotechnical parameter against depth and inspecting parameter values in relation to the soil layering is a central task in geotechnical engineering. Excel-based tools do not offer functionality to easily combine the visualisation of stratigraphy and parameter values. The `SoilProfile` class in `Groundhog` allows easy encoding and manipulation of stratigraphic profiles. The example below generates a stratigraphic profile with three layers. A top layer of sand with a thickness of 5m is underlain by a clay layer of 5m thickness. Below, a silt layer of 10m thickness is found. A constant value of total unit weight of the soil in each layer is specified and the undrained shear strength ($S_u$) trend is defined in the clay layer as a linear variation.
-
-```python
-from groundhog.general.soilprofile import SoilProfile
-sp = SoilProfile({
-    'Depth from [m]': [0, 5, 10],
-    'Depth to [m]': [5, 10, 20],
-    'Soil type': ['Sand', 'Clay', 'Silt'],
-    'Total unit weight [kN/m3]': [19, 17, 18],
-    'Su from [kPa]': [np.nan, 50, np.nan],
-    'Su to [kPa]': [np.nan, 75, np.nan]})
-```
-
[…]
