### 2025-03-13 4e60391
CPT liquefaction updates with docstrings
- added LateX equations to the docstrings
- include the equation in the cpt Correlations
- update the soil parameter mapping
