+++ groundhog/general/parameter_mapping.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-__author__ = 'Bruno Stuyts'
+__author__ = "Bruno Stuyts"
 
 # Native Python packages
 
@@ -14,72 +14,86 @@ from pyproj import Geod
 # Project imports
 
 SOIL_PARAMETER_MAPPING = {
-    'qc [MPa]': 'qc',
-    'fs [MPa]': 'fs',
-    'u2 [MPa]': 'u2',
-    'qt [MPa]': 'qt',
-    'ft [MPa]': 'ft',
-    'qnet [MPa]': 'qnet',
-    'Vertical total stress [kPa]': 'sigma_vo',
-    'Vertical effective stress [kPa]': 'sigma_vo_eff',
-    'Effective pressure [kPa]': 'p_eff',
-    'Effective unit weight [kN/m3]': 'gamma_eff',
-    'Total unit weight [kN/m3]': 'gamma_tot',
-    'Unit weight [kN/m3]': 'gamma',
-    'Ic [-]': 'ic',
-    'Dr [-]': 'relative_density',
-    'Gmax [kPa]': 'gmax',
-    'Qt [-]': 'Qt',
-    'Bq [-]': 'Bq',
-    'Fr [%]': 'Fr',
-    'Rf [%]': 'Rf',
-    'K0 [-]': 'k0',
-    'Vs [m/s]': 'Vs',
-    'gamma [kN/m3]': 'gamma',
-    'OCR [-]': 'ocr',
-    'PI [%]': 'pi',
-    'z [m]': 'depth',
-    'Embedded length [m]': 'embedded_length',
-    'Effective friction angle [deg]': 'phi_eff',
-    'Friction angle [deg]': 'phi',
-    'Critical state friction angle [deg]': 'phi_cs',
-    'Cohesion [kPa]': 'cohesion',
-    'Minor principal stress [kPa]': 'sigma_3',
-    'Major principal stress [kPa]': 'sigma_1',
-    'Interface friction angle [deg]': 'interface_friction_angle',
-    'Undrained shear strength [kPa]': 'undrained_shear_strength',
-    'API soil description': 'api_soildescription',
-    'API relative density description': 'api_relativedensity',
-    'Limiting unit skin friction [kPa]': 'fs_lim',
-    'Limiting unit end bearing [kPa]': 'qb_lim',
-    'Tension modifier [-]': 'tension_modifier',
-    'Borehole diameter [mm]': 'borehole_diameter',
-    'Rod length [m]': 'rod_length',
-    'Country': 'country',
-    'Hammer type': 'hammertype',
-    'Hammer release': 'hammerrelease',
-    'Sampler type': 'samplertype',
-    'N [-]': 'N',
-    'N1_60 [-]': 'N1_60',
-    'eta H [%]': 'eta_H',
-    'eta B [-]': 'eta_B',
-    'eta S [-]': 'eta_S',
-    'eta R [-]': 'eta_R',
-    'd50 [mm]': 'd_50',
-    'N60 [-]': 'N_60',
-    'Granular': 'granular',
-    'Void ratio [-]': 'voidratio',
-    'Dry density [kg/m3]': 'dry_density',
-    'Bulk density [kg/m3]': 'bulk_density',
-    'Water density [kg/m3]': 'water_density',
-    'Density [kg/m3]': 'density',
-    'Gs [-]': 'specific_gravity',
-    'S [-]': 'saturation',
-    'Water content [%]': 'water_content'
+    "qc [MPa]": "qc",
[…]
