+++ groundhog/soildynamics/cptliquefaction.py
@@ -24,18 +24,38 @@ FOS_LIQUEFACTION_ERRORRETURN = {"FoS_liq [-]": np.nan}
 @Validator(FOS_LIQUEFACTION, FOS_LIQUEFACTION_ERRORRETURN)
 def fos_liquefaction(sigma_vo, sigma_vo_eff, CRR, CSR, MSF, K_sigma, **kwargs):
     """
-    Calculates the Factor of Safety (FoS) for liquefaction assessment
+    Calculates the Factor of Safety (FoS) for liquefaction assessment. The FoS is a measure of the capacity of the soil to resist liquefaction relative to the demand imposed by seismic loading.
 
-    Args:
-        sigma_vo [kPa]: Total vertical stress at the depth
-        sigma_vo_eff [kPa]: Effective vertical stress at the depth
-        CRR [-]: Cyclic resistance ratio
-        CSR [-]: cyclic stress ratio
-        MSF [-]: Magnitude scaling factor
-        K_sigma [-]: Overburden correction factor
+    The FoS is computed as the ratio of the Cyclic Resistance Ratio (CRR) to the Cyclic Stress Ratio (CSR), adjusted for earthquake magnitude effects (MSF) and overburden stress effects (K_sigma). The FoS is capped at a maximum value of 5 to avoid unrealistic results.
 
-    Returns:
-        FoS_liq [-]: Factor of Safety (FoS) for liquefaction assessment
+    If the effective vertical stress (:math:`\\sigma_v^{\\prime}`) equals the total vertical stress (:math:`\\sigma_v`), the FoS is set to 5, indicating no risk of liquefaction.
+
+    :param sigma_vo: Total vertical stress at the depth of interest (:math:`\\sigma_v`) [:math:`kPa`] - Suggested range: sigma_vo >= 0.0
+    :param sigma_vo_eff: Effective vertical stress at the depth of interest (:math:`\\sigma_v^{\\prime}`) [:math:`kPa`] - Suggested range: sigma_vo_eff >= 0.0
+    :param CRR: Cyclic Resistance Ratio (:math:`CRR`) [:math:`-`] - Suggested range: 0.0 < CRR <= 1.0
+    :param CSR: Cyclic Stress Ratio (:math:`CSR`) [:math:`-`] - Suggested range: CSR >= 0.0
+    :param MSF: Magnitude Scaling Factor (:math:`MSF`) [:math:`-`] - Suggested range: 0.0 < MSF <= 2.0
+    :param K_sigma: Overburden correction factor (:math:`K_{\\sigma}`) [:math:`-`] - Suggested range: 0.0 < K_sigma <= 2.0
+
+    .. math::
+        FoS = \\min \\left( \\frac{CRR}{CSR} \\cdot MSF \\cdot K_{\\sigma}, 5 \\right)
+
+        \\text{If } \\sigma_v^{\\prime} = \\sigma_v, \\text{ then } FoS = 5
+
+    :returns: Dictionary with the following key:
+
+        - 'FoS_liq [-]': Factor of Safety for liquefaction assessment (:math:`FoS`) [:math:`-`]
+
+    Example:
+        >>> result = fos_liquefaction(
+        ...     sigma_vo=100, sigma_vo_eff=80, CRR=0.25, CSR=0.15, MSF=1.2, K_sigma=1.1
+        ... )
+        >>> print(result)
+        {'FoS_liq [-]': 2.2}
+
+    Reference:
+        - Youd, T. L., et al. (2001). Liquefaction Resistance of Soils: Summary Report from the 1996 NCEER and 1998 NCEER/NSF Workshops on Evaluation of Liquefaction Resistance of Soils. Journal of Geotechnical and Geoenvironmental Engineering, 127(10).
[…]

[…]
