### 2025-03-28 #7636 @Saryace
Hi! I am so honored but this topic is outside of my expertise :(, so I must
say no this time.

On Fri, Mar 21, 2025 at 12:21 PM Rachel Wegener ***@***.***>
wrote:

> 👋🏻 @Saryace <https://github.com/Saryace> and @sinanozaydin
> <https://github.com/sinanozaydin>, would you be willing to review this
> submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
