### 2025-04-11 #7636 @rwegener2
👋🏻 @jlarsen-usgs or @dankusanovic, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
