### 2025-06-05 #7636 @odakan
> 👋🏻 [@earthai-tech](https://github.com/earthai-tech) or [@odakan](https://github.com/odakan), would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Dear @rwegener2,

The Groundhog project looks interesting, and I’d be happy to serve as a reviewer.  I’ll do my best to provide useful and constructive feedback

Best regards,
Deniz
