### 2025-06-06 4c8bdd1
cn_capping is included as an optional parameter in behaviourindex_pcpt_robertsonwride function.
