+++ groundhog/siteinvestigation/insitutests/pcpt_correlations.py
@@ -261,6 +261,7 @@ BEHAVIOURINDEX_PCPT_ROBERTSONWRIDE = {
         "min_value": None,
         "max_value": None,
     },
+    "cn_capping": {"type": "float", "min_value": None, "max_value": None},
 }
 
 BEHAVIOURINDEX_PCPT_ROBERTSONWRIDE_ERRORRETURN = {
@@ -289,6 +290,7 @@ def behaviourindex_pcpt_robertsonwride(
     zhang_subtraction=0.15,
     robertsonwride_coefficient1=3.47,
     robertsonwride_coefficient2=1.22,
+    cn_capping=1.7,
     **kwargs
 ):
     """
@@ -314,7 +316,8 @@ def behaviourindex_pcpt_robertsonwride(
     :param robertsonwride_coefficient2: Second coefficient in the equation by Robertson and Wride (:math:``) [:math:`-`] (optional, default= 1.22)
 
     .. math::
-        Q_{tn} = \\frac{q_t - \\sigma_{vo}}{P_a} \\left( \\frac{P_a}{\\sigma_{vo}^{\\prime}} \\right)^n
+        Cn = \min(1.7, \left(\frac{P_a}{\sigma_{vo}^{\prime}}\right)^n)
+        Q_{tn} = \\frac{q_t - \\sigma_{vo}}{P_a} \\cdot Cn
         \\\\
         n = 0.381 \\cdot I_c + 0.05 \\cdot \\frac{\\sigma_{vo}^{\\prime}}{P_a} - 0.15 \\ \\text{where} \\ n \\leq 1
         \\\\
@@ -340,8 +343,16 @@ def behaviourindex_pcpt_robertsonwride(
 
     """
 
-    def Qtn(qt, sigma_vo, sigma_vo_eff, n, pa=0.001 * atmospheric_pressure):
-        return ((qt - 0.001 * sigma_vo) / pa) * ((pa / (0.001 * sigma_vo_eff)) ** n)
+    def Qtn(
+        qt,
+        sigma_vo,
+        sigma_vo_eff,
+        n,
+        cn_capping=cn_capping,
+        pa=0.001 * atmospheric_pressure,
+    ):
+        cn = min(cn_capping, (pa / (0.001 * sigma_vo_eff)) ** n)
+        return ((qt - 0.001 * sigma_vo) / pa) * cn
 
     def Fr(fs, qt, sigma_vo):
         return 100 * fs / (qt - 0.001 * sigma_vo)
@@ -3089,56 +3100,62 @@ def vs_stressdependent_stuyts(
 
 
 DISSIPATION_TEST_TEH = {
-    'ch': {'type': 'float', 'min_value': 0.0, 'max_value': 100.0},
-    'shearmodulus': {'type': 'float', 'min_value': 0.0, 'max_value': 500000.0},
-    'undrained_shear_strength': {'type': 'float', 'min_value': 1.0, 'max_value': 500.0},
-    'u_initial': {'type': 'float', 'min_value': 0.0, 'max_value': 2000.0},
-    'cone_area': {'type': 'float', 'min_value': 2.0, 'max_value': 15.0},
-    'sensor_location': {'type': 'string', 'options': ("u1", "u2"), 'regex': None},
+    "ch": {"type": "float", "min_value": 0.0, "max_value": 100.0},
+    "shearmodulus": {"type": "float", "min_value": 0.0, "max_value": 500000.0},
+    "undrained_shear_strength": {"type": "float", "min_value": 1.0, "max_value": 500.0},
+    "u_initial": {"type": "float", "min_value": 0.0, "max_value": 2000.0},
[…]
