### 2025-06-20 #8355 @rwegener2
> I reviewed the COI policy and wanted to let you know that the author was a reviewer for my own JOSS paper a couple of years ago. I hope that is not a problem.

Thanks for letting me know @asitha-sena! You are correct, that is not a COI so you can continue as a reviewer for this submission.
