### 2025-06-26 #8355 @odakan
Dear @snakesonabrain,

Thank you for clarifying the contributions.

I truly appreciate the groundhog package—it is useful, timely, and well-documented. The accompanying paper is comprehensive and effectively guides both readers and potential users. The examples and tutorial notebooks are well-prepared and function smoothly as intended.

The modular structure of the package, along with its alignment with established engineering workflows, makes it highly usable not only for researchers but also for practitioners in industry. It fills a long-standing gap in open-source tools for geotechnical modeling and analysis.

FURTHER COMMENTS:

1. In a fresh environment, after installing with pip install groundhog, I had to manually install the following additional dependencies to run the examples:

- Potentially missing: scipy, matplotlib, and jinja2.

2. There appears to be a leftover line from the cookiecutter-pypackage template in the groundhog/__init__.py docstring. As shown below:

```
>>> import groundhog
>>> help (groundhog)
Help on package groundhog:

NAME
    groundhog

DESCRIPTION
    geotechnics
    ~~~~~~

    The geotechnics package - a Python package template project that is intended
    to be used as a cookie-cutter for developing new Python packages.

PACKAGE CONTENTS
    __version__
    consolidation (package)
    constitutivemodels (package)
    deepfoundations (package)
    excavations (package)
    general (package)
    shallowfoundations (package)
    siteinvestigation (package)
    soildynamics (package)
    standards (package)

FILE
    c:\users\deniz\pyenv\groundhog_test_env\lib\site-packages\groundhog\__init__.py
```

@rwegener2 – Overall, groundhog is a functional and valuable tool for both geotechnical and earthquake engineering applications. It shows strong potential for both research and practical use, and in my view, it is worthy of consideration for publication in JOSS.

Thank you again, @snakesonabrain and @rwegener2, for introducing me to this excellent package. I would be glad to contribute to the source code if the opportunity arises.

Best regards,
Deniz
