### 2025-06-28 #8355 @asitha-sena
Hi @snakesonabrain,
I'm slowly making my way through the checklist and have found the both the paper and the Groundhog package to be in excellent condition. However, could please check the references for Matplotlib and Plotly in the paper? They are different from the citing guidance given in https://matplotlib.org/stable/project/citing.html and https://plotly.com/chart-studio-help/citations.
