+++ CHANGES.txt
@@ -1,7 +1,4 @@
-v0.14.0, TBD
-    - .set_cv from ConsolidationCalculation did not accept lists. Fixed thanks to Ricardo Rodriguez Plata 
-    - Added a function for creating a dummy SoilProfile object
-    - Added the solution for dissipation tests according to Teh and Houlsby (1991)
+v0.15.0, TBD
     - TODO: Prepare functionality for SPT fence diagram creation
     - TODO: Create a BoreholeLogging object (to log properties like RQD, UCS, Point load, ...)
     - TODO: Add hatching patterns to LogPlot
@@ -10,6 +7,12 @@ v0.14.0, TBD
     - TODO: Implement soil parameter name conventions across groundhog
     - TODO: Add soil type classification according to Schneider et al (2008)
 
+
+v0.14.0, 03/06/2025 - Ping Li's contribution
+    - .set_cv from ConsolidationCalculation did not accept lists. Fixed thanks to Ricardo Rodriguez Plata 
+    - Added a function for creating a dummy SoilProfile object
+    - Added the solution for dissipation tests according to Teh and Houlsby (1991)
+    - Functionality for liquefaction analysis added by Ping Li
 
 v0.13.0, 11/10/2024 - The Pirate release
     - Update of requirements to Numpy >= 2.1
