### 2025-08-18 930460c
Added soil parameter selection demo
