### 2025-08-19 c2abcee
Further CPT processing implementation
