+++ groundhog/siteinvestigation/insitutests/pcpt_processing.py
@@ -19,7 +19,7 @@ import plotly.graph_objs as go
 from plotly.colors import DEFAULT_PLOTLY_COLORS
 
 # Project imports
-from groundhog.general.plotting import plot_with_log, LogPlotMatplotlib, GROUNDHOG_PLOTTING_CONFIG
+from groundhog.general.plotting import plot_with_log, LogPlotMatplotlib, LogPlot, GROUNDHOG_PLOTTING_CONFIG
 from groundhog.general.parameter_mapping import SOIL_PARAMETER_MAPPING, merge_two_dicts, reverse_dict
 from groundhog.siteinvestigation.insitutests.pcpt_correlations import *
 from groundhog.general.soilprofile import SoilProfile, plot_fence_diagram
@@ -996,6 +996,29 @@ class PCPTProcessing(InsituTestProcessing):
 
     # endregion
 
+    # region Layering selection
+    def select_layering(self, default_soiltype="Unknown", default_unitweight=20, stop_threshold=0, secondaryprofile=None, **kwargs):
+        """
+        Creates a dummy soil profile and creates a ``LogPlotMatplotlib`` from it.
+        It then allows the user to click on the plot to select layers.
+        The ``SoilProfile`` with the selected layering is then returned to the user.
+
+        :param default_soiltype: Default soil type to use for the soil profile resulting from the layer selection.
+        :param default_unitweight: Default unit weight to use for the soil profile resulting from the layer selection.
+        :param stop_threshold: qc values which controls ending the layer selection. Click below the threshold to stop the selection.
+        """
+        dummy_profile = SoilProfile({
+            "Depth from [m]": [0, ],
+            "Depth to [m]": [self.max_depth,],
+            "Soil type": [default_soiltype,],
+            "Total unit weight [kN/m3]": [default_unitweight,]
+        })
+        logplot = self.plot_raw_pcpt_withlog_matplotlib(soilprofile=dummy_profile, secondaryprofile=secondaryprofile, **kwargs)
+        logplot.select_layering()
+        return dummy_profile
+
+    # endregion
+
     # region Layer-based processing and correction
 
     def map_properties(self, layer_profile, cone_profile=DEFAULT_CONE_PROPERTIES,
@@ -1353,7 +1376,216 @@ class PCPTProcessing(InsituTestProcessing):
         else:
             fig.show(config=GROUNDHOG_PLOTTING_CONFIG)
 
+    def plot_raw_pcpt_withlog(self, soilprofile, qc_range=(-10, 100), qc_tick=10, fs_range=(0, 1), fs_tick=0.1,
+                      u2_range=(-0.5, 2.5), u2_tick=0.5, z_range=None, z_tick=2, rf_range=(0, 5), rf_tick=0.5,
+                      plot_friction_ratio=False,
+                      plot_height=700, plot_width=1000, return_fig=False,
+                      plot_title=None, latex_titles=True,
+                      fillcolordict={"Sand": 'yellow', "Clay": 'brown', 'Rock': 'grey'}, **kwargs) :
+        """
+        Plots the raw PCPT data using the Plotly package using a ``groundhog`` ``LogPlot``. This generates an interactive plot.
 
+        :param soilprofile: ``SoilProfile`` object to use for the ``LogPlot``
+        :param qc_range: Range for the cone tip resistance (default=(0, 100MPa))
+        :param qc_tick: Tick interval for the cone tip resistance (default=10MPa)
+        :param fs_range: Range for the sleeve friction (default=(0, 1MPa))
+        :param fs_tick: Tick interval for sleeve friction (default=0.1MPa)
+        :param u2_range: Range for the pore pressure at the shoulder (default=(-0.1, 0.5MPa))
+        :param u2_tick: Tick interval for the pore pressure at the shoulder (default=0.05MPa)
[…]
