+++ groundhog/general/plotting.py
@@ -348,15 +348,16 @@ class LogPlotMatplotlib(object):
     Class for planneled plots with a minilog on the side, using the Matplotlib plotting backend
     """
 
-    def __init__(self, soilprofile, no_panels=1, logwidth=0.05,
+    def __init__(self, soilprofile, secondaryprofile=None, no_panels=1, logwidth=0.05,
                  fillcolordict={"Sand": 'yellow', "Clay": 'brown', 'Rock': 'grey', 'Silt': 'green'},
                  hatchpatterns={"Sand": "...", "Clay": '////', 'Rock':'oo', 'Silt': '|||'},
                  soiltypelegend=True, soiltypecolumn='Soil type', edgecolor='black',
-                 figheight=6, plot_layer_transitions=True, showgrid=True,
+                 figwidth=10, figheight=6, plot_layer_transitions=True, showgrid=True,
                  **kwargs):
         """
         Initializes a figure with a minilog on the side.
         :param soilprofile: Soilprofile used for the minilog
+        :param secondaryprofile: A second Soilprofile used as an additional minilog. Note that only the ``SoilProfile`` provided in ``soilprofile`` will get update if layers and parameters are selected.
         :param no_panels: Number of panels
         :param logwidth: Width of the minilog as a percentage of the total width (default=0.05)
         :param fillcolordict: Dictionary with fill colors for each of the soil types. Every unique ``Soil type`` needs to have a corresponding color. Default: ``{"Sand": 'yellow', "Clay": 'brown', 'Rock': 'grey'}``
@@ -370,14 +371,28 @@ class LogPlotMatplotlib(object):
         :param kwargs: Optional keyword arguments for the make_subplots method
         """
         self.soilprofile = soilprofile
+        
+        if secondaryprofile is None:
+            secondary_exists = False
+        else:
+            secondary_exists = True
+        
+        
         self.no_panels = no_panels
         # Determine the panel widths
-        panel_widths = list(map(lambda _x: (1 - logwidth) / no_panels, range(0, no_panels)))
+        if secondary_exists:
+            panel_widths = list(map(lambda _x: (1 - 2 * logwidth) / no_panels, range(0, no_panels)))
+            panel_widths = list(np.append([logwidth, logwidth], panel_widths))
+            self.no_logs = 2
+        else:
+            panel_widths = list(map(lambda _x: (1 - logwidth) / no_panels, range(0, no_panels)))
+            panel_widths = list(np.append(logwidth, panel_widths))
+            self.no_logs = 1
 
-        panel_widths = list(np.append(logwidth, panel_widths))
+        total_panels = no_panels + self.no_logs
 
         # Set up the figure
-        self.fig, self.axes = plt.subplots(1, no_panels + 1, figsize=(4 * no_panels, figheight), sharex=False, sharey=True,
+        self.fig, self.axes = plt.subplots(1, total_panels, figsize=(figwidth, figheight), sharex=False, sharey=True,
                         constrained_layout=False, gridspec_kw={'width_ratios': panel_widths})
         
         self.axes[0].set_ylim([soilprofile.max_depth, soilprofile.min_depth])
@@ -405,6 +420,27 @@ class LogPlotMatplotlib(object):
             self.axes[0].fill(
                 [0.0,0.0,1.0,1.0],[_y0, _y1, _y1, _y0], fill=True, color=_fillcolor,
                 label='_nolegend_', edgecolor=edgecolor, hatch=_hatch)
+        if secondary_exists:
+            for i, row in secondaryprofile.iterrows():
+                try:
+                    _fillcolor = fillcolordict[row[soiltypecolumn]]
[…]
