### 2025-08-25 70f2506
Added clipping depths calculation according to Tian and Lehane
