+++ groundhog/siteinvestigation/insitutests/pcpt_correlations.py
@@ -8,6 +8,7 @@ __author__ = "Bruno Stuyts"
 # 3rd party packages
 import numpy as np
 from scipy.optimize import brentq
+from scipy.interpolate import interp1d
 import warnings
 
 # Project imports
@@ -3370,6 +3371,119 @@ def dissipation_test_teh(
         'T* [-]': _Tstar,
         'Ir [-]': _Ir,
         'Cone radius [m]': _a
+    }
+
+
+CLIPPINGDEPTHS_QC1N_TIANLEHANE = {
+    'qc1NW': {'type': 'float', 'min_value': 0.0, 'max_value': 1000.0},
+    'qc1NS': {'type': 'float', 'min_value': 0.0, 'max_value': 1000.0},
+    'cone_diameter': {'type': 'float', 'min_value': 0.001, 'max_value': 0.1},
+    'tolerance': {'type': 'float', 'min_value': 0.001, 'max_value': 0.999}
+}
+
+CLIPPINGDEPTHS_QC1N_TIANLEHANE_ERRORRETURN = {
+    'r': np.nan,
+    'eta': np.nan,
+    'qc1N0': np.nan,
+    'as': np.nan,
+    'aw': np.nan,
+    'z*W': np.nan,
+    'z*S': np.nan,
+    'qc1N weak': np.nan,
+    'qc1N strong': np.nan,
+    'qc1N weak function': None,
+    'qc1N strong function': None,
+    'z* clipping weak': np.nan,
+    'z* clipping strong': np.nan,
+    'z clipping weak [m]': np.nan,
+    'z clipping strong [m]': np.nan,
+}
+
+@Validator(CLIPPINGDEPTHS_QC1N_TIANLEHANE, CLIPPINGDEPTHS_QC1N_TIANLEHANE_ERRORRETURN)
+def clippingdepths_qc1N_tianlehane(qc1NW, qc1NS, cone_diameter=0.03568, tolerance=0.05):
+    """
+    Calculates the depths where the normalised cone resistance reaches steady values in weak over strong layer systems based on the equations proposed by Tian and Lehane (2025).
+    These depths can be used to filter CPT data which belongs to a layer transition.
+    The equations were developed based on centrifuge and pressure chamber testing with various two-layer systems (denser sand over looser sand, looser sand over denser sand, sand over clay).
+    Note that the equations provided by Tian and Lehane only work for weaker layers (lower normalised cone resistance) overlying stronger layers (higher normalised cone resistance).
+    
+    :param qc1NW: Steady-state normalised cone resistance in the weaker layer (:math:`q_{c1N,W}`) [-] - Suggested range: 0.0 <= qc1NW <= 1000.0
+    :param qc1NS: Steady-state normalised cone resistance in the stronger layer (:math:`q_{c1N,S}`) [-] - Suggested range: 0.0 <= qc1NS <= 1000.0
+    :param cone_diameter: Cone diameter (:math:`d_c`) [m] - Suggested range: 0.001 <= cone_diameter <= 1000.0 (optional, default=0.03568 for a 10cm2 cone)
+    :param tolerance: Defines the multiplier to detect which data needs to be clipped [-] - Suggested range: 0.001 <= tolerance <= 0.999 (optional, default=0.05)
+
+    .. math::
+        q_{c1N}=q_{c1N,0} - \\tanh \\left[ a_w z^* \\right] \\left( q_{c1N,0} - q_{c1N,W} \\right) \\quad \\text{in weak layer} \\\\
[…]
