### 2025-08-31 4d27c5d
Started on pipelines and cables funcs
