+++ groundhog/siteinvestigation/insitutests/pcpt_correlations.py
@@ -38,6 +38,12 @@ PCPT_NORMALISATIONS_ERRORRETURN = {
     "Qt [-]": np.nan,
     "Fr [-]": np.nan,
     "qnet [MPa]": np.nan,
+    "exponent_zhang [-]": np.nan,
+    "Qtn [-]": np.nan,
+    "Fr [%]": np.nan,
+    "Ic [-]": np.nan,
+    "Ic class number [-]": np.nan,
+    "Ic class": None,
 }
 
 
@@ -52,6 +58,14 @@ def pcpt_normalisations(
     cone_area_ratio,
     start_depth=0.0,
     unitweight_water=10.25,
+    atmospheric_pressure=100,
+    ic_min=1.0,
+    ic_max=4.0,
+    zhang_multiplier_1=0.381,
+    zhang_multiplier_2=0.05,
+    zhang_subtraction=0.15,
+    robertsonwride_coefficient1=3.47,
+    robertsonwride_coefficient2=1.22,
     **kwargs
 ):
     """
@@ -74,6 +88,14 @@ def pcpt_normalisations(
     :param cone_area_ratio: Ratio between the cone rod area and the maximum cone area (:math:`a`) [:math:`-`] - Suggested range: 0.0 <= cone_area_ratio <= 1.0
     :param start_depth: Start depth of the test, specify this for a downhole test. Leave at zero for a test starting from surface (:math:`d`) [:math:`m`] - Suggested range: start_depth >= 0.0 (optional, default= 0.0)
     :param unitweight_water: Unit weight of water, default is for seawater (:math:`\\gamma_w`) [:math:`kN/m3`] - Suggested range: 9.0 <= unitweight_water <= 11.0 (optional, default= 10.25)
+    :param atmospheric_pressure: Atmospheric pressure (used for normalisation) (:math:`P_a`) [:math:`kPa`] (optional, default= 100.0)
+    :param ic_min: Minimum value for soil behaviour type index used in the optimisation routine (:math:`I_{c,min}`) [:math:`-`] (optional, default= 1.0)
+    :param ic_max: Maximum value for soil behaviour type index used in the optimisation routine (:math:`I_{c,max}`) [:math:`-`] (optional, default= 4.0)
+    :param zhang_multiplier_1: First multiplier in the equation for exponent n (:math:``) [:math:`-`] (optional, default= 0.381)
+    :param zhang_multiplier_2: Second multiplier in the equation for exponent n (:math:``) [:math:`-`] (optional, default= 0.05)
+    :param zhang_subtraction: Term subtracted in the equation for exponent n (:math:``) [:math:`-`] (optional, default= 0.15)
+    :param robertsonwride_coefficient1: First coefficient in the equation by Robertson and Wride (:math:``) [:math:`-`] (optional, default= 3.47)
+    :param robertsonwride_coefficient2: Second coefficient in the equation by Robertson and Wride (:math:``) [:math:`-`] (optional, default= 1.22)
 
     .. math::
         q_c = q_c^* + d \\cdot a \\cdot \\gamma_w
@@ -90,6 +112,10 @@ def pcpt_normalisations(
 
         Q_t = \\frac{q_t - \\sigma_{vo}}{\\sigma_{vo}^{\\prime}}
 
+        Q_{tn} = \\frac{q_t - \\sigma_{vo}}{P_a} \\left( \\frac{P_a}{\\sigma_{vo}^{\\prime}} \\right)^n
+        
[…]
