+++ groundhog/siteinvestigation/insitutests/pcpt_processing.py
@@ -1110,7 +1110,7 @@ class PCPTProcessing(InsituTestProcessing):
         else:
             warnings.warn("Downhole corrections have already been applied")
 
-    def normalise_pcpt(self, qc_for_rf=False):
+    def normalise_pcpt(self, qc_for_rf=False, calculate_ic=True, **kwargs):
         """
         Carries out the necessary normalisation and correction on PCPT data to allow calculation of derived parameters and soil type classification.
 
@@ -1123,6 +1123,8 @@ class PCPTProcessing(InsituTestProcessing):
         Note that the absence of pore water pressure measurements will lead to NaN values. Reasoning can be used (e.g. presence of rapidly draining layers) to edit the pore pressure data before running this method.
 
         The total sleeve friction is also calculated. If the cross-sectional area of the friction sleeve is equal at the top and bottom of the sleeve, a correction is not required. This is the default behaviour (cross-sectional areas equal to NaN) but the areas can be entered to calculate the total sleeve friction.
+
+        Soil behaviour type index :math:`I_c` and the corrected normalised cone resistance :math:`Q_{tn}` are calculated if ``calculate_ic`` is set to ``True``. For quicker calculation, set this boolean to False.
 
         .. math::
             q_c = q_c^* + d \\cdot a \\cdot \\gamma_w
@@ -1139,6 +1141,10 @@ class PCPTProcessing(InsituTestProcessing):
 
             Q_t = \\frac{q_t - \\sigma_{vo}}{\\sigma_{vo}^{\\prime}}
 
+            Q_{tn} = \\frac{q_t - \\sigma_{vo}}{P_a} \\left( \\frac{P_a}{\\sigma_{vo}^{\\prime}} \\right)^n
+        
+            n = 0.381 \\cdot I_c + 0.05 \\cdot \\frac{\\sigma_{vo}^{\\prime}}{P_a} - 0.15 \\ \\text{where} \\ n \\leq 1
+
             F_r = \\frac{f_s}{q_t - \\sigma_{vo}}
 
             q_{net} = q_t - \\sigma_{vo}
@@ -1150,28 +1156,56 @@ class PCPTProcessing(InsituTestProcessing):
         :return: Supplements the PCPT data (`.data`) with the normalised properties (column keys 'qt [MPa]', 'Delta u2 [MPa]', 'Rf [%]', 'Bq [-]', 'Qt [-]', 'Fr [%]', 'qnet [MPa]', 'ft [MPa]'
         """
         try:
-            self.data['qt [MPa]'] = self.data['qc [MPa]'] + self.data['u2 [MPa]'] * (1 - self.data['area ratio [-]'])
-            self.data['Delta u2 [MPa]'] = self.data['u2 [MPa]'] - 0.001 * self.data["Hydrostatic pressure [kPa]"]
-            if qc_for_rf:
-                self.data['Rf [%]'] = 100 * self.data['fs [MPa]'] / self.data['qc [MPa]']
-            else:
-                self.data['Rf [%]'] = 100 * self.data['fs [MPa]'] / self.data['qt [MPa]']
-            self.data['Bq [-]'] = self.data['Delta u2 [MPa]'] / (
-                    self.data['qt [MPa]'] - 0.001 * self.data["Vertical total stress [kPa]"])
-            self.data['Qt [-]'] = (self.data['qt [MPa]'] - 0.001 * self.data["Vertical total stress [kPa]"]) / (
-                    0.001 * self.data["Vertical effective stress [kPa]"])
-            self.data['Fr [%]'] = 100 * self.data['fs [MPa]'] / (
-                    self.data['qt [MPa]'] - 0.001 * self.data["Vertical total stress [kPa]"])
[…]
