+++ docs/pipelinescables/stability_toplevel.rst
@@ -1,0 +1,10 @@
+Pipeline and cable stability
+================================
+
+
+.. toctree::
+   :maxdepth: 2
+
+   penetration
+
+   
