### 2025-09-03 6926e8d
Added drained pipeline penetration according to DNV
