+++ groundhog/pipelinescables/stability/penetration.py
@@ -10,6 +10,8 @@ import numpy as np
 
 # Project imports
 from groundhog.general.validation import Validator
+from groundhog.shallowfoundations.capacity import nq_frictionangle_sand, ngamma_frictionangle_vesic, \
+        ngamma_frictionangle_meyerhof, ngamma_frictionangle_davisbooker
 
 """
 Implements pipeline penetration formulae
@@ -326,5 +328,91 @@ def embedment_undrained_method2(
     )
 
     return {
+        'Qv [kN/m]': _Qv,
+    }
+
+
+
+EMBEDMENT_DRAINED = {
+    'penetration': {'type': 'float', 'min_value': 0.0, 'max_value': 2.0},
+    'gamma_eff': {'type': 'float', 'min_value': 3.0, 'max_value': 12.0},
+    'phi_eff': {'type': 'float', 'min_value': 15.0, 'max_value': 50.0},
+    'diameter': {'type': 'float', 'min_value': 0.0, 'max_value': 2.0},
+    'Ngamma_theory': {'type': 'string', 'options': ('Vesic', 'Meyerhof', 'DavisBooker'), 'regex': None},
+}
+
+EMBEDMENT_DRAINED_ERRORRETURN = {
+    'Nq [-]': np.nan,
+    'Ngamma [-]': np.nan,
+    'B [m]': np.nan,
+    'z0 [m]': np.nan,
+    'dq [-]': np.nan,
+    'Qv [kN/m]': np.nan,
+}
+
+@Validator(EMBEDMENT_DRAINED, EMBEDMENT_DRAINED_ERRORRETURN)
+def embedment_drained(
+    penetration,gamma_eff,phi_eff,diameter, roughness_factor=0.67,
+    Ngamma_theory='Vesic', **kwargs):
+
+    """
+    Calculates pipeline embedment in drained conditions using bearing capacity theory. The vertical force required to penetrate the pipe to a given embedment is calculated.
+
+    Note that there is no embedment effect as long as the pipeline is within the active Rankine zone.
+
+    Bearing capacity factors are taken in accordance with shallow foundation theory.
+    
+    :param penetration: Pipeline or cable penetration below virgin seabed (:math:`z`) [m] - Suggested range: 0.0 <= penetration <= 2.0
+    :param gamma_eff: Submerged unit weight of the soil (:math:`\\gamma^{\\prime}`) [kN/m3] - Suggested range: 3.0 <= gamma_eff <= 12.0
+    :param phi_eff: Effective friction angle of the soil (:math:`\\varphi^{\\prime}`) [deg] - Suggested range: 15.0 <= phi_eff <= 50.0
+    :param roughness_factor: Pipeline roughness factor where 0 is fully smooth and 1 is fully rough, used when theory for :math:`N_{\\gamma}` is set to ``'DavisBooker'`` (:math:`R_{inter}`) [:math:`-`] - Suggested range: 0.0 <= roughness_factor <= 1.0
+    :param diameter: Pipeline or cable diameter (:math:`D`) [m] - Suggested range: 0.0 <= diameter <= 2.0
+    :param Ngamma_theory: Select the theoretical formulate of bearing capacity factor Ngamma (optional, default= 'Vesic') - Options: ('Vesic', 'Meyerhof', 'DavisBooker')
+    
+    .. math::
[…]
