### 2025-09-05 4d17c67
Automatic depth column for pydov
