+++ CHANGES.txt
@@ -8,6 +8,7 @@ v0.15.0, TBD
     - Added Ic and Qtn calculation to CPT normalisation
     - Added the Schneider et al (2008) chart for plotting CPT data
     - Started added functionality for pipeline and cable penetration
+    - Added automatic detection of depth column for pydov CPTs
     - TODO: Prepare functionality for SPT fence diagram creation
     - TODO: Create a BoreholeLogging object (to log properties like RQD, UCS, Point load, ...)
     - TODO: Add hatching patterns to LogPlot
