+++ groundhog/general/plotting.py
@@ -176,7 +176,7 @@ class LogPlot(object):
     Class for planneled plots with a minilog on the side.
     """
 
-    def __init__(self, soilprofile, no_panels=1, logwidth=0.05,
+    def __init__(self, soilprofile, secondaryprofile=None, no_panels=1, logwidth=0.05,
                  fillcolordict={"Sand": 'yellow', "Clay": 'brown', 'Rock': 'grey'},
                  soiltypelegend=True,
                  soiltypecolumn="Soil type",
@@ -185,6 +185,7 @@ class LogPlot(object):
         """
         Initializes a figure with a minilog on the side.
         :param soilprofile: Soilprofile used for the minilog
+        :param secondaryprofile: A second Soilprofile used as an additional minilog. Note that only the ``SoilProfile`` provided in ``soilprofile`` will get update if layers and parameters are selected.
         :param no_panels: Number of panels
         :param logwidth: Width of the minilog as a ratio to the total width of the figure (default=0.05)
         :param fillcolordict: Dictionary with fill colors for each of the soil types. Every unique ``Soil type`` needs to have a corresponding color. Default: ``{"Sand": 'yellow', "Clay": 'brown', 'Rock': 'grey'}``
@@ -195,15 +196,30 @@ class LogPlot(object):
         """
         self.soilprofile = soilprofile
 
-        # Determine the panel widths
-        panel_widths = list(map(lambda _x: (1 - logwidth) / no_panels, range(0, no_panels)))
+        if secondaryprofile is None:
+            secondary_exists = False
+        else:
+            secondary_exists = True
 
-        panel_widths = list(np.append(logwidth, panel_widths))
+        self.no_panels = no_panels
+        # Determine the panel widths
+        if secondary_exists:
+            
+            panel_widths = list(map(lambda _x: (1 - 2 * logwidth) / no_panels, range(0, no_panels)))
+            panel_widths = list(np.append([logwidth, logwidth], panel_widths))
+            self.no_logs = 2
+        else:
+            panel_widths = list(map(lambda _x: (1 - logwidth) / no_panels, range(0, no_panels)))
+            panel_widths = list(np.append(logwidth, panel_widths))
+            self.no_logs = 1
+
+        total_panels = no_panels + self.no_logs
 
         # Set up the figure
         self.fig = subplots.make_subplots(
-            rows=1, cols=no_panels+1, column_widths=panel_widths, shared_yaxes=True,
+            rows=1, cols=total_panels, column_widths=panel_widths, shared_yaxes=True,
             print_grid=False, **kwargs)
+        
         self.fig['layout']['yaxis1'].update(range=(soilprofile.max_depth, soilprofile.min_depth))
 
         # Create rectangles for the log plot
@@ -217,6 +233,17 @@ class LogPlot(object):
             _y1 = row[self.soilprofile.depth_to_col]
             _layers.append(
                 dict(type='rect', xref='x1', yref='y', x0=0, y0=_y0, x1=1, y1=_y1, fillcolor=_fillcolor, opacity=1, line_width=line_width))
+
+        if secondary_exists:
+            for i, row in secondaryprofile.iterrows():
+                try:
+                    _fillcolor = fillcolordict[row[soiltypecolumn]]
+                except:
+                    _fillcolor = DEFAULT_PLOTLY_COLORS[i % 10]
+                _y0 = row[secondaryprofile.depth_from_col]
+                _y1 = row[secondaryprofile.depth_to_col]
+                _layers.append(
+                    dict(type='rect', xref='x2', yref='y', x0=0, y0=_y0, x1=1, y1=_y1, fillcolor=_fillcolor, opacity=1, line_width=line_width))
 
         for _soiltype in soilprofile[soiltypecolumn].unique():
[…]
