### 2025-09-05 6c479c1
Added layering retrieval from Flanders geological model
