### 2025-09-05 3f697b8
Added lay amplification formula
