### 2025-09-05 0ca066b
Unit test for lay_touchdown_factor
