### 2025-09-06 052a0f6
Maximum number of characters for DOV geological units
