+++ CHANGES.txt
@@ -11,6 +11,7 @@ v0.15.0, TBD
     - Added automatic detection of depth column for pydov CPTs
     - Added secondary soilprofile to LogPlot too
     - Added functionality for retrieving stratigraphic information from DOV (Flanders 3D geological model)
+    - Added function to retrieve BRO CPT data based on BRO CPT ID
     - TODO: Prepare functionality for SPT fence diagram creation
     - TODO: Create a BoreholeLogging object (to log properties like RQD, UCS, Point load, ...)
     - TODO: Add hatching patterns to LogPlot
