+++ groundhog/general/soilprofile.py
@@ -608,6 +608,41 @@ class SoilProfile(pd.DataFrame):
         self.dropna(subset=(self.depth_from_col, self.depth_to_col), inplace=True)
         self.reset_index(drop=True, inplace=True)
 
+    def merge_soiltypes(self):
+        """
+        Merge adjacent layers with same soil type
+        """
+        layer_no = 0
+        for i, row in self.iterrows():
+            self.loc[i, 'Layer no'] = layer_no
+            if i > 0:
+                if row['Soil type'] != self.loc[i-1, "Soil type"]:
+                    layer_no += 1
+                self.loc[i, 'Layer no'] = layer_no
+        
+        for _layer in self['Layer no'].unique():
+            _layer_data = self[self['Layer no'] == _layer]
+            _layer_top = _layer_data[self.depth_from_col].min()
+            _layer_bottom = _layer_data[self.depth_to_col].max()
+            values_dict = dict()
+            for _param in self.numerical_soil_parameters():
+                if self.check_linear_variation(_param):
+                    values_dict[_param] = [
+                        self.loc[_layer_data.index[0], _param.replace(' [', ' from [')],
+                        self.loc[_layer_data.index[-1], _param.replace(' [', ' to [')],
+                    ]
+            
+            if _layer_data.__len__() == 1:
+                pass
+            else:
+                self.loc[_layer_data.index[0], "Depth to [m]"] = _layer_bottom
+                for _key, _value in values_dict.items():
+                    self.loc[_layer_data.index[0], _key.replace(' [', ' from [')] = _value[0]
+                    self.loc[_layer_data.index[0], _key.replace(' [', ' to [')] = _value[1]
+                self.drop(index=_layer_data.index[1:], inplace=True)
+        self.drop('Layer no', axis=1, inplace=True)
+        self.reset_index(drop=True, inplace=True)
+
     def remove_parameter(self, parameter):
         """
         Removes a soil parameter from the dataframe
@@ -1326,7 +1361,7 @@ class CalculationGrid(object):
         x = np.insert(second_array_x, np.arange(len(first_array_x)), first_array_x)
         return z, x
 
-# region DOV functionality
+# region DOV and BRO functionality
 def retrieve_geological_profile_dov(x, y, model="g3dv3_L", namecutoff=30, **kwargs):
     """
     Retrieves a geological profile at a location with given coordinates (x,y Lambert L72).
@@ -1378,5 +1413,457 @@ def retrieve_geological_profile_dov(x, y, model="g3dv3_L", namecutoff=30, **kwar
 
     except Exception as err:
         print("Error during retrieval of geological layering - %s" % str(err))
+
+BRO_FILLCOLORS_PLOTLY_LITHOK = {
+    'antropogeen': 'rgb(200,200,200)',
+    'organisch materiaal (veen)': 'rgb(157,78,64)',
+    'klei': 'rgb(0,146,0)',
+    'kleiig zand, zandige klei en leem': 'rgb(194,207,92)',
[…]
