+++ groundhog/siteinvestigation/insitutests/pcpt_processing.py
@@ -1526,41 +1526,43 @@ class PCPTProcessing(InsituTestProcessing):
         
         cptplot = LogPlot(
             soilprofile=soilprofile, no_panels=3, fillcolordict=fillcolordict, **kwargs)
-        cptplot.add_trace(
-            x=self.data['qc [MPa]'],
-            z=self.data['z [m]'],
-            name='qc',
-            panel_no=1,
-            showlegend=False,
-            line=dict(color='black', width=1)
-        )
-        cptplot.add_trace(
-            x=self.data['fs [MPa]'],
-            z=self.data['z [m]'],
-            name='fs',
-            panel_no=2,
-            showlegend=False,
-            line=dict(color='black', width=1)
-        )
+        for _push in self.data["Push"].unique():
+            push_data = self.data[self.data["Push"] == _push]
+            cptplot.add_trace(
+                x=push_data['qc [MPa]'],
+                z=push_data['z [m]'],
+                name='qc',
+                panel_no=1,
+                showlegend=False,
+                line=dict(color='black', width=1)
+            )
+            cptplot.add_trace(
+                x=push_data['fs [MPa]'],
+                z=push_data['z [m]'],
+                name='fs',
+                panel_no=2,
+                showlegend=False,
+                line=dict(color='black', width=1)
+            )
 
-        if plot_friction_ratio:
-            cptplot.add_trace(
-                x=self.data['Rf [%]'],
-                z=self.data['z [m]'],
-                name='Rf',
-                panel_no=3,
-                showlegend=False,
-                line=dict(color='black', width=1)
-            )
-        else:
-            cptplot.add_trace(
-                x=self.data['u2 [MPa]'],
-                z=self.data['z [m]'],
-                name='u2',
-                panel_no=3,
-                showlegend=False,
-                line=dict(color='black', width=1)
-            )
+            if plot_friction_ratio:
+                cptplot.add_trace(
+                    x=push_data['Rf [%]'],
+                    z=push_data['z [m]'],
+                    name='Rf',
+                    panel_no=3,
+                    showlegend=False,
+                    line=dict(color='black', width=1)
+                )
+            else:
+                cptplot.add_trace(
+                    x=push_data['u2 [MPa]'],
+                    z=push_data['z [m]'],
+                    name='u2',
+                    panel_no=3,
+                    showlegend=False,
+                    line=dict(color='black', width=1)
+                )
         cptplot.set_xaxis(title=qc_axis_title, panel_no=1, range=qc_range, dtick=qc_tick)
         cptplot.set_xaxis(title=fs_axis_title, panel_no=2, range=fs_range, dtick=fs_tick)
         if plot_friction_ratio:
@@ -1630,41 +1632,43 @@ class PCPTProcessing(InsituTestProcessing):
         
         cptplot = LogPlotMatplotlib(
             soilprofile=soilprofile, no_panels=3, fillcolordict=fillcolordict, **kwargs)
-        cptplot.add_trace(
-            x=self.data['qc [MPa]'],
-            z=self.data['z [m]'],
-            name='qc',
-            panel_no=1,
-            showlegend=False,
-            c='black'
-        )
-        cptplot.add_trace(
-            x=self.data['fs [MPa]'],
-            z=self.data['z [m]'],
-            name='fs',
-            panel_no=2,
-            showlegend=False,
-            c='black'
-        )
+        for _push in self.data["Push"].unique():
+            push_data = self.data[self.data["Push"] == _push]
+            cptplot.add_trace(
+                x=push_data['qc [MPa]'],
+                z=push_data['z [m]'],
+                name='qc',
+                panel_no=1,
+                showlegend=False,
+                c='black'
+            )
+            cptplot.add_trace(
+                x=push_data['fs [MPa]'],
[…]
