### 2025-09-19 a4a54b8
Added routine for (Robertson) soil type selection
