### 2025-09-19 86f1310
Update of unit tests following ic calculation update
