### 2025-09-19 ce3d954
Added Cn capping to normalisation routine
