### 2025-09-19 3ba4eff
Addressed JOSS reviewer 2 comments
