### 2025-10-03 #8355 @asitha-sena
Hi @rwegener2 and @snakesonabrain
It doesn't seem like the reference for Matplotlib has been fixed yet. Please check https://matplotlib.org/stable/project/citing.html for guidance.

I have no other concerns.
