+++ groundhog/__version__.py
@@ -7,7 +7,7 @@ Information about the current version of the py-package-template package.
 
 __title__ = 'groundhog'
 __description__ = 'Groundhog - A general-purpose geotechnical package'
-__version__ = '0.15.0'
+__version__ = '0.16.0'
 __author__ = 'Bruno Stuyts'
 __author_email__ = 'bruno@pro-found.be'
 __license__ = 'GNU GPLv3'
