### 2025-10-03 #8355 @snakesonabrain
The Matplotlib reference change was not synched with the online branch. Should be good now!
