### 2025-10-04 #8355 @snakesonabrain
@asitha-sena I generated the paper again and checked the reference to Matplotlib. It is now OK. Thanks a lot for the review!
