### 2025-10-10 2f08031
Added bending stiffness calculations for soilmix walls
