+++ groundhog/excavations/soilmix.py
@@ -1,0 +1,242 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+__author__ = "Bruno Stuyts"
+
+# Native Python packages
+
+# 3rd party packages
+import numpy as np
+
+# Project imports
+from groundhog.general.validation import Validator
+
+
+BENDINGSTIFFNESS_SOILMIX_METHOD1 = {
+    'moment_inertia_reinforcement': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'modulus_soilmix': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'height_soilmix': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'reinforcement_offset': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'height_reinforcement': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'flange_thickness': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'connection_thickness': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'flange_width': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'modulus_reinforcement': {'type': 'float', 'min_value': 0.0, 'max_value': 300000000.0},
+    'participating_width': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'tensile_strength_soilmix': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+}
+
+BENDINGSTIFFNESS_SOILMIX_METHOD1_ERRORRETURN = {
+    'n [-]': np.nan,
+    'Ism [m4]': np.nan,
+    'EI-1 [kNm2]': np.nan,
+    'Mcr [kNm]': np.nan,
+    'c1 [m]': np.nan,
+    'c2 [m]': np.nan,
+    'd [m]': np.nan,
+    'c1b [m]': np.nan,
+    'hw [m]': np.nan,
+    'Af [m2]': np.nan,
+    'rho [-]': np.nan,
+    'xi,e [-]': np.nan,
+    'xe [m]': np.nan,
+    'I2 [m4]': np.nan,
+    'EI-2 [kNm2]': np.nan,
+    'EI [kNm2]': np.nan,
+    'EI-eff/m [kNm2/m]': np.nan
+}
+
+@Validator(BENDINGSTIFFNESS_SOILMIX_METHOD1, BENDINGSTIFFNESS_SOILMIX_METHOD1_ERRORRETURN)
+def bendingstiffness_soilmix_method1(
+    moment_inertia_reinforcement, modulus_soilmix, height_soilmix, reinforcement_offset, height_reinforcement,
+    flange_thickness, connection_thickness, flange_width,
+    modulus_reinforcement=210000000.0, participating_width=np.nan, tensile_strength_soilmix=np.nan, **kwargs):
+
+    """
+    Calculates the bending stiffness of a soilmix wall (for use in geotechnical retaining wall calculations) using the combined stiffness of reinforcements and the soilmix material itself.
+    The stiffness is calculated as the average of the cracked and uncracked stiffness of the material according to §5.4.2.3 if EN 1992-1-1. The formulae contain simplications that result in a small error of 1 to 3%.
+    Note that all dimensions are in meters and units of force are in kN.
+    
+    :param moment_inertia_reinforcement: Moment of inertia of the reinforcement (:math:`I_a`) [m4] - Suggested range: moment_inertia_reinforcement >= 0.0
[…]
