+++ docs/excavations/excavations_toplevel.rst
@@ -7,4 +7,4 @@ Excavations
 
    basic
 
-   
+   soilmix
