### 2025-10-29 7ea06b3
Added concrete section and elastic solutions
