+++ groundhog/shallowfoundations/stressdistribution.py
@@ -348,3 +348,144 @@ def stresses_rectangle(
         'delta sigma y [kPa]': _delta_sigma_y,
         'delta tau zx [kPa]': _delta_tau_zx,
     }
+
+
+STRESSES_LINELOAD_RETAININGWALL = {
+    'lineload': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'toe_depth': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'horizontal_offset': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'depth': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+}
+
+STRESSES_LINELOAD_RETAININGWALL_ERRORRETURN = {
+    'delta sigma x [kPa]': np.nan,
+    'delta P x [kN/m]': np.nan,
+}
+
+@Validator(STRESSES_LINELOAD_RETAININGWALL, STRESSES_LINELOAD_RETAININGWALL_ERRORRETURN)
+def stresses_lineload_retainingwall(
+    lineload, toe_depth, horizontal_offset, depth, **kwargs):
+
+    """
+    Calculates the elastic stress increase due to a line load (infinitely long out of plane) next to a buried earth-retaining structure.
+    
+    :param lineload: Magnitude of the applied line load (:math:`Q`) [kN/m] - Suggested range: lineload >= 0.0
+    :param toe_depth: Depth of the toe of the retaining wall (:math:`H_0`) [m] - Suggested range: toe_depth >= 0.0
+    :param horizontal_offset: Offset between the line load and the retaining structure (:math:`a H_0`) [m] - Suggested range: horizontal_offset >= 0.0
+    :param depth: Depth considered for the calculation (cannot be deeper than the toe depth) (:math:`b H_0`) [m] - Suggested range: depth >= 0.0
+    
+    .. math::
+        \\Delta \\sigma_x = \\frac{4 Q a^2 b}{\\pi H_0 \\left( a^2 + b^2 \\right)^2}
+        
+        \\Delta P_x = \\frac{2 Q}{\\pi \\left( a^2 + 1 \\right)}
+    
+    :returns: Dictionary with the following keys:
+        
+        - 'delta sigma x [kPa]': Increase of horizontal stress (:math:`\\Delta \\sigma_x`)  [kPa]
+        - 'delta P x [kN/m]': Increase of horizontal force (:math:`\\Delta P_x`)  [kN/m]
+    
+    Reference - Budhu (2011). Soil mechanics and foundation engineering
+
+    """
+    if depth > toe_depth:
+        raise ValueError("Depth cannot exceed the toe depth")
+    _a = horizontal_offset / toe_depth
+    _b = depth / toe_depth
+    _delta_sigma_x = (4 * lineload * (_a ** 2) * _b) / (np.pi * toe_depth * (_a ** 2 + _b ** 2) ** 2)
+    _delta_P_x = (2 * lineload) / (np.pi * (_a ** 2 + 1))
+
+    return {
+        'delta sigma x [kPa]': _delta_sigma_x,
+        'delta P x [kN/m]': _delta_P_x,
+    }
+
+
+STRESSES_STRIPLOAD_RETAININGWALL = {
+    'imposedstress': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'width': {'type': 'float', 'min_value': 0.0, 'max_value': None},
+    'offset': {'type': 'float', 'min_value': 0.0, 'max_value': None},
[…]
