### 2025-10-29 962cf2d
Included modulus ratio
