### 2025-11-05 f1e54c2
Added LCPC method
