+++ docs/piles/lcpc.rst
@@ -1,0 +1,7 @@
+LCPC pile resistance calculations
+====================================
+
+.. autoclass:: groundhog.deepfoundations.axialcapacity.lcpc.LCPCAxcapCalculation
+    :members:
+
+    .. automethod:: __init__
