+++ docs/piles/piles_toplevel.rst
@@ -15,6 +15,8 @@ Pile calculations
 
    koppejan
 
+   lcpc
+
    settlement
 
    cavityexpansion
