### 2025-11-16 08be500
Chin-kondler method, group effect and doc fixes
