+++ groundhog/deepfoundations/lateralresponse/lateral.py
@@ -7,6 +7,8 @@ __author__ = "Bruno Stuyts"
 
 # 3rd party packages
 import numpy as np
+from plotly import tools, subplots
+import plotly.graph_objs as go
 
 # Project imports
 from groundhog.general.validation import Validator
@@ -111,4 +113,152 @@ def reinforced_circularsection_inertia(diameter, modulus_ratio, n_bars, offset, 
         'Rebar inertia [m4]': _total_rebar_inertia,
         'Ic [m4]': _Ic,
         'I combined [m4]': _Ic + _total_rebar_inertia
+    }
+
+PILEGROUPEFFECT_REESEVANIMPE = {
+    'pile_x': {'type': 'list', 'elementtype': 'float', 'order': None, 'unique': False, 'empty_allowed': False},
+    'pile_x': {'type': 'list', 'elementtype': 'float', 'order': None, 'unique': False, 'empty_allowed': False},
+    'pile_diameters': {'type': 'list', 'elementtype': 'float', 'order': None, 'unique': False, 'empty_allowed': False},
+    'load_x': {"type": "float", "min_value": None, "max_value": None},
+    'load_y': {"type": "float", "min_value": None, "max_value": None}
+}
+
+PILEGROUPEFFECT_REESEVANIMPE_ERRORRETURN = {
+    'efficiency_matrix': None,
+    'efficiencies': None,
+    'pile_fig': None
+}
+
+@Validator(PILEGROUPEFFECT_REESEVANIMPE, PILEGROUPEFFECT_REESEVANIMPE_ERRORRETURN)
+def pilegroupeffect_reesevanimpe(pile_x, pile_y, pile_diameters, load_x, load_y, show_fig=True, plot_height=600, plot_width=400, **kwargs):
+    """
+    When piles are arranged in a group, they influence one another and the lateral reaction for a given displacement can be less than that for a single pile.
+    Reese and Van Impe suggest a method for calculating the efficiency of each pile. A distinction is made between in-line leading piles, in-line trailing piles and side-by-side piles.
+    Based on the direction of loading and inter-pile distance, a reduction factor (p-multiplier) is calculated for each pile pair.
+    The efficiency factors for each pile pair are multiplied to provide the overall efficiency for the pile considered.
+    In this function, the centers of the piles are defined using their (X,Y) coordinates.
+    The direction of loading is defined using the X- and Y-component of the loading vector. Note that the magnitude of this loading vector (norm) does not play a role in the calculation.
+    The diameter of the pile is also required to determine the normalised pile spacing.
+    For piles which is neither perfectly inline or side-by-side, the inline and side-by-side efficiencies are combined using the angle to the loading direction.
+
+    :param pile_x: List of X-coordinates of the pile centers (:math:`x`) [m]
+    :param pile_y: List of Y-coordinates of the pile centers (:math:`y`) [m]
+    :param pile_diameters: List of pile diameters (:math:`y`) [m]
+    :param load_x: X-component of the load vector (:math:`x_{\\text{load}}`)
+    :param load_y: Y-component of the load vector (:math:`x_{\\text{load}}`)
+    :param show_fig: Boolean determining whether the figures for the Chin-Kondler construction need to be plotted (default behaviour) or returned in the output dictionary.
+
+    .. math::
[…]
