+++ tests/deepfoundations/test_lateral.py
@@ -16,4 +16,14 @@ class Test_lateral(unittest.TestCase):
         result_2 = lateral.reinforced_circularsection_inertia(diameter=0.42, modulus_ratio=7, n_bars=2, offset=0.5, rebar_diameter=0.016, maximum_resistance=False)
         self.assertAlmostEqual(result_2["Start angle [deg]"], 0, 1)
         self.assertGreaterEqual(result_1["Rebar inertia [m4]"], result_2["Rebar inertia [m4]"])
+
+    def test_pilegroupeffect_reesevanimpe(self):
+        result = lateral.pilegroupeffect_reesevanimpe(
+            pile_x=[0, 1.58, 0, 1.58, 0, 1.58, 0, 1.58],
+            pile_y=[0, 0, 1.08, 1.08, 2.16, 2.16, 3.18, 3.18],
+            pile_diameters=[0.42, 0.42, 0.42, 0.42, 0.42, 0.42, 0.42, 0.42],
+            load_x=2, load_y=2, show_fig=False
+        )
+        self.assertAlmostEqual(result['efficiencies'][0], 0.404, 3)
+        self.assertAlmostEqual(result['efficiencies'][-1], 0.707, 3)
         
