+++ docs/piles/piletesting.rst
@@ -1,0 +1,5 @@
+Pile testing functionality
+=================================================
+
+.. automodule:: groundhog.deepfoundations.axialcapacity.piletesting
+    :members:
