+++ docs/piles/piles_toplevel.rst
@@ -19,6 +19,10 @@ Pile calculations
 
    settlement
 
+   lateral
+
    cavityexpansion
 
    negativeskinfriction
+
+   piletesting
