+++ docs/piles/lateral.rst
@@ -1,0 +1,5 @@
+Pile lateral behaviour
+====================================
+
+.. automodule:: groundhog.deepfoundations.lateralresponse.lateral
+    :members:

[…]
