### 2026-01-05 #8355 @rwegener2
Thanks @snakesonabrain!

We generally recommend that the paper and the zenodo archive have the same title and author lists. Do you have a specific motivation for using different titles and author lists for the zenodo archive and the JOSS paper?
