### 2026-01-23 #8355 @kthyng
@snakesonabrain Paper looks good except please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "north sea".
