### 2026-01-23 #8355 @snakesonabrain
@kthyng I changed the capitalization. I changed everything to lower case for consistency except for first letters of first words and names of geographical entities.
Thanks for looking through the paper.
