### 2026-01-23 #8355 @kthyng
> [@snakesonabrain](https://github.com/snakesonabrain)
>
> > I had previously created a Zenodo entry for the groundhog code repository and the Zenodo archive is for the code repo, not for the paper itself. Should I create a separate Zenodo entry for the paper?
>
> It's fine using this Zenodo entry without the paper as long as it is pointing to the version that was reviewed here, which looks to be the case. [@rwegener2](https://github.com/rwegener2) Was asking if you want to update the metadata (author list and title) for that Zenodo archive to match this submission title and paper title to create a cohesive set of archives, since they all go together, but it is your choice.

Can you address this comment?
